+++ src/command/args/vmaf.rs
@@ -9,6 +9,18 @@ pub struct Vmaf {
     /// See https://ffmpeg.org/ffmpeg-filters.html#libvmaf.
     #[clap(long = "vmaf", parse(from_str = parse_vmaf_arg))]
     pub vmaf_args: Vec<Arc<str>>,
+
+    /// Video resolution width to use in VMAF analysis. If set, video streams will be bicupic
+    /// scaled to this width during VMAF analysis. By default automatically set based on the
+    /// model and input video resolution. Setting to `0` disables any such scaling.
+    ///
+    /// Default automatic behaviour:
+    /// * Video width <  1728 => scale to 1920 (1080p)
+    /// * Video width >= 1728 => no scaling
+    ///
+    /// Scaling happens after any input/reference vfilters.
+    #[clap(long)]
+    pub vmaf_width: Option<u32>,
 }
 
 fn parse_vmaf_arg(arg: &str) -> Arc<str> {
@@ -26,15 +38,20 @@ impl Vmaf {
         let mut lavfi = args.join(":");
         lavfi.insert_str(0, "libvmaf=");
 
-        let scale_1080p = !args.iter().any(|a| a.contains("model"))
-            && distorted_width.map_or(false, |w| w < 1728);
+        let vmaf_w = match (self.vmaf_width, distorted_width) {
+            (None, Some(w)) => match VmafModel::from_args(&args) {
+                // upscale small resolutions to 1k for use with the 1k model
+                VmafModel::Vmaf1K if w < 1728 => Some(1920),
+                _ => None,
+            },
+            (w, _) => w.filter(|w| *w != 0 && Some(*w) != distorted_width),
+        };
 
-        if scale_1080p {
-            // upscale both streams to 1080p, as this resolution is what the default
-            // vmaf model is designed to work with
+        if let Some(w) = vmaf_w {
+            // scale both streams to the vmaf width
             lavfi.insert_str(
                 0,
-                "[0:v]scale=1920:-1:flags=bicubic[dis];[1:v]scale=1920:-1:flags=bicubic[ref];[dis][ref]",
+                &format!("[0:v]scale={w}:-1:flags=bicubic[dis];[1:v]scale={w}:-1:flags=bicubic[ref];[dis][ref]"),
             );
         }
 
@@ -42,17 +59,42 @@ impl Vmaf {
     }
 }
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+enum VmafModel {
+    /// Default 1080p model.
+    Vmaf1K,
+    /// Some other user specified model.
+    Custom,
+}
+
+impl VmafModel {
+    fn from_args(args: &[Arc<str>]) -> Self {
+        let using_custom_model = args
+            .iter()
+            .filter(|v| !v.ends_with("version=vmaf_v0.6.1"))
+            .any(|v| v.contains("model"));
+        match using_custom_model {
+            true => Self::Custom,
+            false => Self::Vmaf1K,
+        }
+    }
+}
+
 #[test]
 fn vmaf_lavfi() {
     let vmaf = Vmaf {
         vmaf_args: vec!["n_threads=5".into(), "n_subsample=4".into()],
+        vmaf_width: None,
     };
     assert_eq!(vmaf.ffmpeg_lavfi(None), "libvmaf=n_threads=5:n_subsample=4");
 }
 
 #[test]
 fn vmaf_lavfi_default() {
-    let vmaf = Vmaf { vmaf_args: vec![] };
+    let vmaf = Vmaf {
+        vmaf_args: vec![],
+        vmaf_width: None,
+    };
     let expected = format!("libvmaf=n_threads={}", num_cpus::get());
     assert_eq!(vmaf.ffmpeg_lavfi(None), expected);
 }
@@ -61,6 +103,7 @@ fn vmaf_lavfi_default() {
 fn vmaf_lavfi_include_n_threads() {
     let vmaf = Vmaf {
         vmaf_args: vec!["log_path=output.xml".into()],
+        vmaf_width: None,
     };
     let expected = format!("libvmaf=log_path=output.xml:n_threads={}", num_cpus::get());
     assert_eq!(vmaf.ffmpeg_lavfi(None), expected);
@@ -71,6 +114,7 @@ fn vmaf_lavfi_include_n_threads() {
 fn vmaf_lavfi_small_width() {
     let vmaf = Vmaf {
         vmaf_args: vec!["n_threads=5".into(), "n_subsample=4".into()],
+        vmaf_width: None,
     };
     assert_eq!(
         vmaf.ffmpeg_lavfi(Some(1280)),
@@ -80,7 +124,7 @@ fn vmaf_lavfi_small_width() {
     );
 }
 
-/// If user has overriden the model, don't presume to scale
+/// If user has overriden the model, don't default a vmaf width
 #[test]
 fn vmaf_lavfi_small_width_custom_model() {
     let vmaf = Vmaf {
@@ -89,6 +133,7 @@ fn vmaf_lavfi_small_width_custom_model() {
             "n_threads=5".into(),
             "n_subsample=4".into(),
         ],
+        vmaf_width: None,
     };
     assert_eq!(
         vmaf.ffmpeg_lavfi(Some(1280)),
@@ -97,9 +142,29 @@ fn vmaf_lavfi_small_width_custom_model() {
 }
 
 #[test]
+fn vmaf_lavfi_custom_model_and_width() {
+    let vmaf = Vmaf {
+        vmaf_args: vec![
+            "model=version=foo".into(),
+            "n_threads=5".into(),
+            "n_subsample=4".into(),
+        ],
+        // if specified just do it
+        vmaf_width: Some(123),
+    };
+    assert_eq!(
+        vmaf.ffmpeg_lavfi(Some(1280)),
+        "[0:v]scale=123:-1:flags=bicubic[dis];\
+        [1:v]scale=123:-1:flags=bicubic[ref];\
+        [dis][ref]libvmaf=model=version=foo:n_threads=5:n_subsample=4"
+    );
+}
+
+#[test]
 fn vmaf_lavfi_1080p() {
     let vmaf = Vmaf {
         vmaf_args: vec!["n_threads=5".into(), "n_subsample=4".into()],
+        vmaf_width: None,
     };
     assert_eq!(
         vmaf.ffmpeg_lavfi(Some(1920)),